refactor: DistanceResponse, FareResponse 삭제
